[+] ALF-3503 - Настройка наборов данных в отчетах (мультиорганизации и контекст пользователя) [~] изменения при деплое/удалении отчетов - подчистка зависающих в системе файлов с шаблонами [+] проверка в механизме справочников на то, что удаляемый узел уже был удален (policy) или помечен, как удаляемый (при удалении родителя)
git-svn-id: http://svn.datateh.ru/alfrescoecm/Business%20Platform/trunk@10092 697f8a5b-bb4b-4f4c-ab90-ff1ad96ece97

+++ lecm-dictionary/repo/src/main/java/ru/it/lecm/dictionary/beans/XMLImportBeanImpl.java
@@ -192,7 +192,9 @@ public class XMLImportBeanImpl implements XMLImportBean {
             if (updateMode.isRewriteChildren()) {
                 List<ChildAssociationRef> childAssocs = nodeService.getChildAssocs(current);
                 for (ChildAssociationRef childAssoc : childAssocs) {
-                    nodeService.deleteNode(childAssoc.getChildRef());
+                    if (nodeService.exists(childAssoc.getChildRef()) && !nodeService.hasAspect(childAssoc.getChildRef(), ContentModel.ASPECT_PENDING_DELETE)) {
+                        nodeService.deleteNode(childAssoc.getChildRef());
+                    }
                 }
             }
             readItems(xmlr, current, updateMode);
